% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshroom-package.R
\docType{package}
\name{ggshroom-package}
\alias{ggshroom}
\alias{ggshroom-package}
\title{ggshroom package}
\description{
Fungal and fungus-like icons for ggplot2 plots.
}
\author{
\strong{Maintainer}: Geoff Shroom \email{geoff@localhost}

}
\keyword{internal}
